+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9.0] - 2020-09-24
+
+### Breaking
+
+- Fixed inability to playback recordings if the length of any of the
+  secrets provided differ at playback time from recording time.  The
+  length of redactions provided for playback must be derived from the
+  length of the original secret that was redacted.
+
 ## [0.8.1] - 2020-09-24
 
 ### Changed